Finished Part 12. Skipping Parts 13 and 14 for now.
Part 12's TODO list:
- Start implementing plan limitations
- Make it possible to subscribe to a plan

As stated in the title for the commit and in more detail in my log, I am skipping parts 13 and 14 because it involves something called "Stripe" which, after skimming through those two videos, seems to involve setting up the plans to actually work with payment. Since I don't need to do this right now and am not interested in creating an account to do that, I am skipping those two videos/parts for now and instead working on part 15 and maybe part 16 depending on what I can and can't do because of my skipping of parts 13 and 14.