Finished Part 15 and this project for now (see desc for a bit more info).
TODO list from Part 15:
- Make it possible to cancel a plan
- Show plan end date in team page (*skipped*)
- Make it possible to delete a lead
- Make it possible to delete a client
- Only show "leads, clients, and team" in menu when authenticated

This project for now is complete. I skipped Parts 13, 14, and 16 of the tutorial, at least for now. Maybe I will come back in the future to finish those parts, but I want to start and finish a few more projects for GitHub and my resume before my break is over.